Handle all checkboxes being unchecked.
Use a hidden field to tell the difference between the page being loaded
for the first time and the form being submitted with all the checkboxes
unchecked.
I'm not totally happy with this solution but given checkboxes submit
nothing at all when they are unchecked I can't think of a better
approach right now. Improvements welcome!

The finding support tests now pass.