Update slimmer version and use explicit layout
In Slimmer 9.0.0 the default layout changed from `wrapper` to
`core_layout`, which uses a modern grid and components. We're trying
to move away from `wrapper`, as it contains a lot of very broad styling
and selectors, making it difficult to update without breaking other
apps.

This app can't be moved to `core_layout` yet as it relies on features
that aren't available there yet;

- related sidebar
- breadcrumbs
- Some copy formatting

Once those are available, most likely as components, this app can be
moved away from `wrapper`. In the mean time this gives us an up to
date version of slimmer, and is explicit about using an older layout.