Send valid content to the publishing-api
The content schemas require a `details` hash, and `format` has been
replaced by `document_type` and `schema_name`.

This should have been picked up by the branch tests on
govuk-content-schemas, but it seems we don't run the tests for this app
whenever the schema changes. We'll fix that in a separate commit.